Refactor namespaces for consistency and clarity
