+++ src/Grains.Impl/States/FusionSession.cs
@@ -1,4 +1,4 @@
-namespace Argon.Api.Grains.Persistence.States;
+namespace Grains.Impl.States;
 
 using MemoryPack;
 using Orleans;
