+++ src/Models/DTO/SfuPermissionFlags.cs
@@ -1,4 +1,4 @@
-﻿namespace Argon.Api.Features.Sfu;
+﻿namespace Models.DTO;
 
 [Flags]
 public enum SfuPermissionFlags
